Add tests to voter logic
